A commit of my driver implementation for the K64F microcontroller
A commit of my driver implementation for the K64F microcontroller, using the Kinetis SDK V2 FLEXCAN driver, and FreeRTOS's critical section functions.

This has been thoroughly tested, but is not the most efficient implementation possible, as a number of hardware features in the FLEXCAN module are instead implemented in the driver software. It is still however efficient enough to saturate a 1MHz CANbus with minimal CPU time.